api/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
aiosqlite
python-dotenv